requirements.txt
python-telegram-bot[job-queue]>=20.0
pymongo>=4.0